backend/requirements.txt
# Flask核心框架
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0

# 数据库相关
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
SQLAlchemy>=2.0.27

# 密码加密
Flask-Bcrypt==1.0.1
bcrypt==4.0.1

# 环境变量管理
python-dotenv==1.0.0

# 表单处理和安全
Flask-WTF==1.2.2
WTForms==3.0.1

# 开发和测试工具
pytest==7.4.2
pytest-flask==1.2.0
flake8==6.1.0
black==23.9.1

# 其他依赖
Werkzeug>=3.0.0
Jinja2>=3.1.2
click>=8.1.7
MarkupSafe>=2.1.3
itsdangerous>=2.1.2